refactor: separate common note elements to NoteCard
