+++ src/components/Note/types.ts
@@ -1,0 +1,5 @@
+import { INote } from '@store/slices/notesSlice/types';
+
+export type INoteComponent = {
+  baseNote: INote;
+};
